# order-service/docs/swagger.yaml
definitions:
  handler.CreateOrderItemRequest:
    properties:
      price:
        example: 10.99
        type: number
      product_id:
        example: prod-001
        type: string
      quantity:
        example: 2
        type: integer
    required:
    - price
    - product_id
    - quantity
    type: object
  handler.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.CreateOrderItemRequest'
        type: array
      user_id:
        example: 123e4567-e89b-12d3-a456-426655440000
        type: string
    required:
    - items
    - user_id
    type: object
  handler.OrderItemResponse:
    properties:
      id:
        type: string
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  handler.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/handler.OrderItemResponse'
        type: array
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Get a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.OrderResponse'
            type: array
      summary: List all orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with the provided items
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.OrderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OrderResponse'
      summary: Get an order by ID
      tags:
      - orders
swagger: "2.0"